ex_ichannel::bindings

Function __ic_forward_reply_to

pub unsafe extern "C" fn __ic_forward_reply_to(
    ic: *mut ichannel_t,
    slot: u64,
    cmd: i32,
    res: *const c_void,
    exn: *const c_void,
)

\brief Bounce an IOP answer to reply to another slot.

This function may be used to forward an answer to another slot when implementing a manual proxy. It saves the reply data packing.

WARNINGS:

this function MUST be used before leaving the RPC callback scope.

you cannot use this function to forward a webservices answer; this is not implemented.

be really careful because this function suppose that the answer has been left unmodified since its reception. If you want to modify it before the forwarding, then don’t use this function and use instead ic_reply_p/ic_throw_p. If you try to do this, in the best scenario your changes will be ignored, in the worst you will have a crash…

Here an example of how to use this function: RPC_IMPL(pkg, foo, bar) { CHECK_ID_OK(arg->id);

 // manual proxy
 ic_msg_t *imsg = ic_msg_new(sizeof(uint64_t));
 *(uint64_t *)imsg->priv = slot;
 ic_query_p(remote_ic, imsg, pkg, foo, bar, arg);

}

RPC_CB(pkg, foo, bar) { uint64_t origin_slot = *(uint64_t *)msg->priv;

 // automatic and efficient answer forwarding
 __ic_forward_reply_to(ic, origin_slot, status, res, exn);

}

\param[in] ic the ichannel_t the “thing” we proxy comes from. \param[in] slot the slot of the query we’re answering to. \param[in] cmd the received answer status parameter. \param[in] res the received answer result parameter. \param[in] exn the received answer exception parameter.