ex_ichannel::bindings

Function farch_get_filename

pub unsafe extern "C" fn farch_get_filename(
    file: *const farch_entry_t,
    name_outbuf: *mut i8,
) -> *mut i8

Get the filename of this entry.

This function can be used to iterate on names entries: entries are stored in an array terminating by an empty entry (name.len == 0).

for (const farch_entry_t *entry = entries,
         char __buf[FARCH_MAX_FILENAME];
     name = farch_get_filename(entry, __buf);
     entry++)

\param[in] file a farch entry. \param[in|out] name a buffer to write the name – should contains FARCH_MAX_FILENAME bytes. \return NULL if there is no entry or if an error occurs, name otherwise.