ex_ichannel::bindings

Function iop_get_cvar

pub unsafe extern "C" fn iop_get_cvar(
    obj: *const c_void,
    name: lstr_t,
) -> *const iop_value_t

Gets the value of a class variable (static field).

This takes a class instance pointer and a class variable name, and returns a pointer of the value of this class variable for the given object type.

If the wanted static field does not exist in the given class, this function will return NULL.

It also assumes that the given pointer is a valid pointer on a valid class instance. If not, it will probably crash…

\param[in] obj Pointer on a class instance. \param[in] name Name of the wanted class variable.