ex_ichannel::bindings

Function iop_msort_desc

pub unsafe extern "C" fn iop_msort_desc(
    iop_env_ctx: *const iop_env_ctx_t,
    st: *const iop_struct_t,
    vec: *mut c_void,
    len: i32,
    params: *const qv_iop_sort_t,
    err: *mut sb_t,
) -> i32

Sort a vector of IOP as iop_sort, but on multiple fields.

\param[in] iop_env_ctx The current IOP environment context. \param[in] st The IOP structure definition (__s). \param[in] vec The array to sort \see iop_sort. \param[in] len Length of the array \param[in] params All the path to sort on, the array will be sorted on the first element, and in case of equality, on the seconde one, and so on. \see iop_sort for field path syntax and flags desc. \param[out] err In case of error, the error description.