ex_ichannel::bindings

Function mem_ring_reset

pub unsafe extern "C" fn mem_ring_reset(arg1: *mut mem_pool_t)

Force clean the ring pool from empty memory blocks.

Adapted from mem_stack_pool_reset().

Only keeps the current frame so as the one that fits the best the current needs regarding the mean allocation size (see code for details).

Should be called when the pool is idle: won’t do anything if called while there are still some frames not released.

XXX The clean-up is supposed to be automatically triggered every 64K frame releases, using this function to force it means that you have a clear understanding of what you are doing.