ex_ichannel::bindings

Function mpsc_queue_looks_empty

pub unsafe extern "C" fn mpsc_queue_looks_empty(
    q: *mut mpsc_queue_t,
) -> bool

\brief tells whether an mpsc queue looks empty or not.

This should be only used from the only mpsc consumer thread.

The function may return false positives but no false negatives: the consumer thread is the sole consumer, hence if it sees a non empty queue the queue can’t be freed without it knowing about it.

If the function returns true though, some other thread may enqueue something at the same time and the function will return a false positive, hence the name “the queue looks empty but maybe it isn’t”.

\param[in] q the queue \returns true if the queue looks empty, false else.