ex_ichannel::bindings

Function thr_schedule

pub unsafe extern "C" fn thr_schedule(job: *mut thr_job_t)

\brief Schedule one job.

By default, jobs are queued in a per-thread deque queue. The deque allows:

non blocking job queuing

non blocking job dequeuing if there is more than one job in queue

CAS-guarded job dequeuing if there is one job in queue

CAS-guarded job stealing from other threads.

Jobs order is NOT preserved, as the local thread will take the last jobs queued first, and stealing threads will take the first ones first.

The job life-cycle is the responsibility of the caller.

Note that each thread can only queue up to a fixed number of jobs. Any job in excess is run immediately instead of being queued.

In the fast paths, job queuing and dequeuing are very fast (below 40-50 cycles probably). Stealing a job in the fast path is pretty fast too, though searching for a stealable job isn’t (atm).

It’s a good idea to be sure that the job you queue is long enough so that the stealing costs are amortized (especially if the job structure is malloc()ed).