ex_ichannel::bindings

Struct ichannel_t

#[repr(C)]
pub struct ichannel_t {

    pub name: lstr_t,
    pub su: sockunion_t,
    pub remote_addr: lstr_t,
    pub impl_: *const qm_ic_cbs_t,
    pub iop_env: *const iop_env_t,
    pub on_event: Option<unsafe extern "C" fn(*mut ichannel_t, ic_event_t)>,
    pub on_wipe: Option<unsafe extern "C" fn(*mut ichannel_t)>,
    pub owner: *mut *mut ichannel_t,
    pub priv_: *mut c_void,
    pub addr_fail_log_level: i32,
    pub dns_ctx: *mut dns_resolv_ctx_t,
    pub peer: *mut c_void,
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 3]>,
    pub id: u32,
    pub protocol: i32,
    pub nextslot: u32,
    pub elh: *mut ev_t,
    pub timer: *mut ev_t,
    pub desc: *const iop_rpc_t,
    pub cmd: i32,
    pub priority: ev_priority_t,
    pub wa_soft_timer: *mut ev_t,
    pub wa_soft: i32,
    pub wa_hard: i32,
    pub peer_version: u16,
    pub cb_level: i16,
    pub retry_delay: i32,
    pub on_creds: Option<unsafe extern "C" fn(*mut ichannel_t, *const ic_creds_t) -> i32>,
    pub queries: qm_ic_msg_t,
    pub iov_list: htlist_t,
    pub msg_list: htlist_t,
    pub last_normal_prio_msg: *mut htnode_t,
    pub local_async_queries: htlist_t,
    pub local_async_replies: htlist_t,
    pub el_before: *mut ev_t,
    pub current_fd: i32,
    pub pending: i32,
    pub queue_len: i32,
    pub iov_total_len: i32,
    pub fds: qv_i32_t,
    pub iov: qv_iovec_t,
    pub rbuf: sb_t,
    pub ssl: *mut ssl_st,
    pub client_addr: lstr_t,
    pub pending_max: i32,

}

Fields
name: lstr_t

Optional name of the IChannel.

This name is used for exploitation purposes only.

su: sockunion_t

Resolved address.

Filled with the resolved client or server address. For a client IChannel, this field can be pre-filled if the address resolution is done before; otherwise, the remote_addr field must be filled.

remote_addr: lstr_t

Client IChannels: non-resolved remote address.

If provided, this address will be resolved and used for each connection attempt; otherwise, the resolution must be done before and provided through the su field. This field is only used for TCP client IChannels.

impl_: *const qm_ic_cbs_t

Map of the RPC implemented for this IC.

iop_env: *const iop_env_t

Mandatory IOP environment for the IC.

on_event: Option<unsafe extern "C" fn(*mut ichannel_t, ic_event_t)>

Mandatory callback to get notified about the connection state (connected, disconnected, no activity, …).

on_wipe: Option<unsafe extern "C" fn(*mut ichannel_t)>

Optional callback called before wiping the IC.

owner: *mut *mut ichannel_t

Content set to NULL on deletion

priv_: *mut c_void

User private data.

addr_fail_log_level: i32

Log level of the address resolution when it fails. This log level ensures the process can eventually stop if the address resolution cannot be achieved.

dns_ctx: *mut dns_resolv_ctx_t

Context linked to an address resolution currently in progress.

peer: *mut c_void

User field to identify the peer.

_bitfield_1: __BindgenBitfieldUnit<[u8; 3]>id: u32

IChannel unique ID.

The IChannel unique ID is put inside the messages slots so the IChannel can be retrieved for async replies.

protocol: i32

Transport layer protocol.

Default is 0.

nextslot: u32

Next slot ID to try for messages slots allocation.

elh: *mut ev_t

Event loop handle of the connection or wake object on asynchronous address resolution.

timer: *mut ev_t

Timer of the activity watcher.

desc: *const iop_rpc_t

Description of the current unpacked RPC.

cmd: i32

Command value of the current unpacked structure.

priority: ev_priority_t

Event loop priority of the IChannel.

wa_soft_timer: *mut ev_t

Soft timer of the activity watcher.

wa_soft: i32

Timeout to be notified when there is no activity.

wa_hard: i32

Timeout to close the connection when there is no activity.

peer_version: u16

Protocol version of the remote peer.

cb_level: i16

Account the number of entered callbacks.

retry_delay: i32

Delay before a reconnection attempt (ms).

Default is 1000ms.

on_creds: Option<unsafe extern "C" fn(*mut ichannel_t, *const ic_creds_t) -> i32>

Callback optionally set by ic_swawn() to retrieve the ic_creds_t of the incoming Unix connection.

queries: qm_ic_msg_t

Map of queries awaiting for an answer.

iov_list: htlist_t

List of messages awaiting to be sent, using an IOV.

msg_list: htlist_t

List of messages awaiting to be sent.

last_normal_prio_msg: *mut htnode_t

Last message of msg_list having the priority NORMAL.

local_async_queries: htlist_tlocal_async_replies: htlist_tel_before: *mut ev_t

Handle used to defer some actions.

Can be used to defer actions that cannot be executed synchronously, like scheduling local async queries and replies, or perform the automatic deletion of the IChannel.

current_fd: i32

Used to store the current file description exchanged on a Unix socket. See ic_get_fd.

pending: i32

Number of pending queries (for peak warning).

queue_len: i32

Length of the queue of queries, without accounting canceled queries.

iov_total_len: i32

Length of the IOV buffer filled for the next write.

fds: qv_i32_t

Internal stack of file descriptors wrapped in the current message.

iov: qv_iovec_t

Vector of IOV stacking for the next write.

rbuf: sb_t

Internal reading buffer.

ssl: *mut ssl_st

TLS context, if any.

client_addr: lstr_t

Server IChannels: client IP address.

This field is the IP address of the client connected to the IChannel. This field should not be read directly, ic_get_client_addr should be used instead.

pending_max: i32

Used in non-release builds to warning about sudden traffic increase that could require investigation.

Implementations

impl ichannel_t

pub fn do_el_unref(&self) -> bool

pub fn set_do_el_unref(&mut self, val: bool)

pub unsafe fn do_el_unref_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_do_el_unref_raw(this: *mut ichannel_t, val: bool)

pub fn auto_reconn(&self) -> bool

pub fn set_auto_reconn(&mut self, val: bool)

pub unsafe fn auto_reconn_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_auto_reconn_raw(this: *mut ichannel_t, val: bool)

pub fn no_autodel(&self) -> bool

pub fn set_no_autodel(&mut self, val: bool)

pub unsafe fn no_autodel_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_no_autodel_raw(this: *mut ichannel_t, val: bool)

pub fn is_trusted(&self) -> bool

pub fn set_is_trusted(&mut self, val: bool)

pub unsafe fn is_trusted_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_trusted_raw(this: *mut ichannel_t, val: bool)

pub fn is_public(&self) -> bool

pub fn set_is_public(&mut self, val: bool)

pub unsafe fn is_public_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_public_raw(this: *mut ichannel_t, val: bool)

pub fn tls_required(&self) -> bool

pub fn set_tls_required(&mut self, val: bool)

pub unsafe fn tls_required_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_tls_required_raw(this: *mut ichannel_t, val: bool)

pub fn is_closing(&self) -> bool

pub fn set_is_closing(&mut self, val: bool)

pub unsafe fn is_closing_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_closing_raw(this: *mut ichannel_t, val: bool)

pub fn is_spawned(&self) -> bool

pub fn set_is_spawned(&mut self, val: bool)

pub unsafe fn is_spawned_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_spawned_raw(this: *mut ichannel_t, val: bool)

pub fn is_seqpacket(&self) -> bool

pub fn set_is_seqpacket(&mut self, val: bool)

pub unsafe fn is_seqpacket_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_seqpacket_raw(this: *mut ichannel_t, val: bool)

pub fn is_unix(&self) -> bool

pub fn set_is_unix(&mut self, val: bool)

pub unsafe fn is_unix_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_unix_raw(this: *mut ichannel_t, val: bool)

pub fn is_wiped(&self) -> bool

pub fn set_is_wiped(&mut self, val: bool)

pub unsafe fn is_wiped_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_wiped_raw(this: *mut ichannel_t, val: bool)

pub fn cancel_guard(&self) -> bool

pub fn set_cancel_guard(&mut self, val: bool)

pub unsafe fn cancel_guard_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_cancel_guard_raw(this: *mut ichannel_t, val: bool)

pub fn queuable(&self) -> bool

pub fn set_queuable(&mut self, val: bool)

pub unsafe fn queuable_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_queuable_raw(this: *mut ichannel_t, val: bool)

pub fn is_local(&self) -> bool

pub fn set_is_local(&mut self, val: bool)

pub unsafe fn is_local_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_local_raw(this: *mut ichannel_t, val: bool)

pub fn is_local_async(&self) -> bool

pub fn set_is_local_async(&mut self, val: bool)

pub unsafe fn is_local_async_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_local_async_raw(this: *mut ichannel_t, val: bool)

pub fn fd_overflow(&self) -> bool

pub fn set_fd_overflow(&mut self, val: bool)

pub unsafe fn fd_overflow_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_fd_overflow_raw(this: *mut ichannel_t, val: bool)

pub fn hdr_checked(&self) -> bool

pub fn set_hdr_checked(&mut self, val: bool)

pub unsafe fn hdr_checked_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_hdr_checked_raw(this: *mut ichannel_t, val: bool)

pub fn is_connected(&self) -> bool

pub fn set_is_connected(&mut self, val: bool)

pub unsafe fn is_connected_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_is_connected_raw(this: *mut ichannel_t, val: bool)

pub fn no_user_version_check(&self) -> bool

pub fn set_no_user_version_check(&mut self, val: bool)

pub unsafe fn no_user_version_check_raw(this: *const ichannel_t) -> bool

pub unsafe fn set_no_user_version_check_raw(this: *mut ichannel_t, val: bool)

pub fn new_bitfield_1( do_el_unref: bool, auto_reconn: bool, no_autodel: bool, is_trusted: bool, is_public: bool, tls_required: bool, is_closing: bool, is_spawned: bool, is_seqpacket: bool, is_unix: bool, is_wiped: bool, cancel_guard: bool, queuable: bool, is_local: bool, is_local_async: bool, fd_overflow: bool, hdr_checked: bool, is_connected: bool, no_user_version_check: bool, ) -> __BindgenBitfieldUnit<[u8; 3]>