ex_ichannel::bindings

Type Alias pointer_formatter_f

pub type pointer_formatter_f = Option<unsafe extern "C" fn(i32, *const c_void, *mut _IO_FILE, *mut i8, usize) -> isize>;

Pointer Formatter function type.

A pointer formatter is called when, for a formatter f, the format %pf is encountered in the format string. The formatter is then called with the parameters passed as argument of that particular format as well as the modifier.

see \p formatter_f description for more details.

\param[in] modifier The modifier that triggered the call \param[in] val The pointer received as parameter. \param[out] stream If the formatter is expected to write on a stream, this is the destination stream. \param[out] buf If the formatter is expected to write in a buffer, this is the destination buffer. \param[out] buf_len The size of the buffer.

\return The size required to format the \p val (not including the trailing zero) or -1 in case of error.

Aliased Type

pub enum pointer_formatter_f {
    None,
    Some(unsafe extern "C" fn(i32, *const c_void, *mut _IO_FILE, *mut i8, usize) -> isize),
}

Variants

None

No value.

Some(unsafe extern "C" fn(i32, *const c_void, *mut _IO_FILE, *mut i8, usize) -> isize)

Some value of type T.